Disable agroal only after the ready checks
